対応済みボタンを移動
comments.vueからapp/views/talks/show.html.slimに移動
