+++ app/javascript/comments.js
@@ -9,7 +9,6 @@ document.addEventListener('DOMContentLoaded', () => {
     const commentableId = comments.getAttribute('data-commentable-id')
     const commentableType = comments.getAttribute('data-commentable-type')
     const currentUserId = Number(comments.getAttribute('data-current-user-id'))
-    const initialCompleted = comments.getAttribute('data-is-completed')
     const currentUser = window.currentUser
     new Vue({
       store,
@@ -19,8 +18,7 @@ document.addEventListener('DOMContentLoaded', () => {
             commentableId: commentableId,
             commentableType: commentableType,
             currentUserId: currentUserId,
-            currentUser: currentUser,
-            initialActionCompleted: initialCompleted === 'true'
+            currentUser: currentUser
           }
         })
     }).$mount(selector)
