+++ app/views/talks/show.html.slim
@@ -51,7 +51,9 @@
                             image_class: 'page-content-members__user-icon'
                           = link_to user_path(member.id), class: 'a-user-name' do
                             = member.login_name
-        #js-comments(data-commentable-id="#{@talk.id}" data-commentable-type='Talk' data-current-user-id="#{current_user.id}" data-is-completed="#{@talk.action_completed}")
+        - if current_user.admin?
+          div(data-vue="ActionCompleted" data-vue-commentable-id:number="#{@talk.id}" data-vue-is-initial-completed:boolean="#{@talk.action_completed}")
+        #js-comments(data-commentable-id="#{@talk.id}" data-commentable-type='Talk' data-current-user-id="#{current_user.id}")
 
       - if current_user.admin?
         .col-xl-5.col-xs-12
